+++ Sources/BinaryStore/Box+StringArray.swift
@@ -8,12 +8,12 @@ extension BinaryStore.Box {
     //   offsetWidth: offset
     //   byteWidth: count of data bytes
     public func getStringArray(index: Int, offsetWidth: BinaryStore.BitWidth, byteWidth: BinaryStore.BitWidth = .bit32, arrayWidth: BinaryStore.BitWidth, stringWidth: BinaryStore.BitWidth = .bit8, encoding: String.Encoding = .utf8) -> [String] {
-        let sz: UInt64 = getInt(offset: index + offsetWidth.rawValue, intWidth: byteWidth)
+        let sz: Int = getInt(offset: index + offsetWidth.rawValue, intWidth: byteWidth)
         if sz == 0 {
             return []
         }
-        let off: UInt64 = getInt(offset: index, intWidth: offsetWidth)
-        return getStringArray(offset: Int(off), arrayWidth: arrayWidth, stringWidth: stringWidth, encoding: encoding)
+        let off: Int = getInt(offset: index, intWidth: offsetWidth)
+        return getStringArray(offset: off, arrayWidth: arrayWidth, stringWidth: stringWidth, encoding: encoding)
     }
